status codes added

+++ src/student/student.service.ts
@@ -1,35 +1,44 @@
-import { Injectable} from '@nestjs/common';
+import { Injectable, HttpException, HttpStatus} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Student } from 'src/typeorm/entities/Students';
 import { Repository } from 'typeorm';
 import { CreateStudentDto, UpdateStudentDto } from './dto/student.dto';
-
 @Injectable()
 export class StudentService {
     constructor(@InjectRepository(Student) private studentRepository:Repository<Student>){}
     // createStudent(createStudentDto:CreateStudentDto){
     //     return createStudentDto;
     // }
-    fetchStudents(id):Promise<Student>{
-        return this.studentRepository.findOneBy({id});
+    async fetchStudents(id):Promise<Student>{
+        if(await this.studentRepository.countBy({id})>0){
+            return this.studentRepository.findOneBy({id});
+        }
     }
     createStudent(createStudentDetails:CreateStudentDto){
         const newStudent=this.studentRepository.create({
-            ...createStudentDetails,createdAt:new Date(),
+            ...createStudentDetails,createdAt:new Date(),modifiedAt:new Date()
         });
-        return this.studentRepository.save(newStudent);
+        this.studentRepository.save(newStudent);
+        
+        return this.fetchStudents(newStudent.id);
     }
-    updateStudent(id:number,updateStudentData:UpdateStudentDto){
-        if(updateStudentData){
-        return this.studentRepository.update({id},{...updateStudentData,modifiedAt:new Date()});
-        // return this.studentRepository.findOneBy({id});
-        // this.studentRepository.findOneBy({id});
+    async updateStudent(id:number,updateStudentData:UpdateStudentDto){
+        if(await this.studentRepository.countBy({id})>0){
+            let row=await this.studentRepository.findOneBy({id});
+            return this.studentRepository.save({...row,...updateStudentData});
         }
         else{
-            this.studentRepository.update({id},{...updateStudentData,modifiedAt:null});
+            throw new HttpException('no such row found', HttpStatus.FORBIDDEN);
+
         }
     }
-    deleteStudent(id:number){
-        return this.studentRepository.delete({id});
+    async deleteStudent(id:number){
+        if(await this.studentRepository.countBy({id})>0){
+            this.studentRepository.delete({id});
+            return 0
+        }
+        else{
+            throw new HttpException('no such row found', HttpStatus.FORBIDDEN);
+        }
     }
 }
